requirements.txt
numpy
matplotlib
streamlit
plotly
pandas
scipy
Pillow
